requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
build==1.2.2.post1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.17
distro==1.9.0
Django==4.2
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
django-mptt
multidict==6.1.0
netaddr==1.3.0
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
pip-tools==7.4.1
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyproject_hooks==1.2.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
